+++ android-unity-wrapper/unitywrapper/src/main/java/com/appsflyer/unity/AppsFlyerAndroidWrapper.java
@@ -4,8 +4,11 @@ package com.appsflyer.unity;
 import androidx.annotation.NonNull;
 
 import com.appsflyer.AFLogger;
+import com.appsflyer.AFPurchaseDetails;
+import com.appsflyer.AFPurchaseType;
 import com.appsflyer.AppsFlyerConsent;
 import com.appsflyer.AppsFlyerConversionListener;
+import com.appsflyer.AppsFlyerInAppPurchaseValidationCallback;
 import com.appsflyer.AppsFlyerInAppPurchaseValidatorListener;
 import com.appsflyer.AppsFlyerLib;
 import com.appsflyer.AppsFlyerProperties;
@@ -29,6 +32,8 @@ public class AppsFlyerAndroidWrapper {
 
     private static final String VALIDATE_CALLBACK = "didFinishValidateReceipt";
     private static final String VALIDATE_ERROR_CALLBACK = "didFinishValidateReceiptWithError";
+    private static final String VALIDATE_AND_LOG_V2_CALLBACK = "onValidateAndLogComplete";
+    private static final String VALIDATE_AND_LOG_V2__ERROR_CALLBACK = "onValidateAndLogFailure";
     private static final String GCD_CALLBACK = "onConversionDataSuccess";
     private static final String GCD_ERROR_CALLBACK = "onConversionDataFail";
     private static final String OAOA_CALLBACK = "onAppOpenAttribution";
@@ -38,7 +43,7 @@ public class AppsFlyerAndroidWrapper {
     private static final String ON_DEEPLINKING = "onDeepLinking";
     private static final String START_REQUEST_CALLBACK = "requestResponseReceived";
     private static final String IN_APP_RESPONSE_CALLBACK = "inAppResponseReceived";
-    private static final String PLUGIN_VERSION = "6.14.0";
+    private static final String PLUGIN_VERSION = "6.14.3";
     private static final long DDL_TIMEOUT_DEFAULT = 3000;
     private static AppsFlyerConversionListener conversionListener;
     private static String devkey = "";
@@ -249,6 +254,16 @@ public class AppsFlyerAndroidWrapper {
         }
     }
 
+    public static void validateAndTrackInAppPurchaseV2(int purchaseType, String purchaseToken, String productId, String price, String currency, HashMap<String, String> additionalParameters, final String objectName) {
+        AFPurchaseType type = purchaseType == 0 ? AFPurchaseType.SUBSCRIPTION : AFPurchaseType.ONE_TIME_PURCHASE;
+        AFPurchaseDetails details = new AFPurchaseDetails(type, purchaseToken, productId, price, currency);
+
+        if (objectName != null){
+            AppsFlyerInAppPurchaseValidationCallback listener = initInAppPurchaseValidatorV2Listener(objectName);
+            AppsFlyerLib.getInstance().validateAndLogInAppPurchase(details, additionalParameters, listener);
+        }
+    }
+
     public static boolean isTrackingStopped() {
         return AppsFlyerLib.getInstance().isStopped();
     }
@@ -345,6 +360,26 @@ public class AppsFlyerAndroidWrapper {
                 }
             }
         });
+    }
+
+    public static AppsFlyerInAppPurchaseValidationCallback initInAppPurchaseValidatorV2Listener(final String objectName) {
+        return new AppsFlyerInAppPurchaseValidationCallback() {
+            @Override
+            public void onInAppPurchaseValidationFinished(@NonNull Map<String, ?> map) {
+                if (objectName != null) {
+                    JSONObject jsonObject = new JSONObject(map);
+                    UnityPlayer.UnitySendMessage(objectName, VALIDATE_AND_LOG_V2_CALLBACK, jsonObject.toString());
+                }
+            }
+
+            @Override
+            public void onInAppPurchaseValidationError(@NonNull Map<String, ?> map) {
+                if (objectName != null) {
+                    JSONObject jsonObject = new JSONObject(map);
+                    UnityPlayer.UnitySendMessage(objectName, VALIDATE_AND_LOG_V2__ERROR_CALLBACK, jsonObject.toString());
+                }
+            }
+        };
     }
 
     public static void handlePushNotifications(){